refactor : FrontController 의 역할을 하는 VoucherManagementController 클래스 구현
